backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
aiosqlite==0.19.0
asgiref==3.7.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
Pillow==12.0.0
python-dotenv==1.0.1
reportlab==4.0.9
sentence-transformers==2.5.1
openai
faiss-cpu
numpy